Improve ./bacardi format command. (#111)
We already have ./bacardi format command to format our source code
automatically. But it works only when the change is in stage.
For example, if someone already commits some changes locally, there is
no opportunity to format the source code. So, this patch increases the
range of applying format. (HEAD~20)

Also, this patch is fixing some format errors.

ISSUE=none